+++ src/PostType/ManualEvent.php
@@ -220,6 +220,15 @@ class ManualEvent {
                 'label'        => 'Päättymisajankohta',
                 'instructions' => '',
             ],
+            'recurring_event'    => [
+                'label'        => 'Toistuva tapahtuma',
+                'instructions' => '',
+            ],
+            'recurring_dates'    => [
+                'label'        => 'Toistuvan tapahtuman ajankohdat',
+                'button_label' => 'Lisää ajankohta',
+                'instructions' => '',
+            ],
             'location'           => [
                 'label'       => 'Sijainti',
                 'name'        => [
@@ -319,6 +328,32 @@ class ManualEvent {
                 ->set_display_format( 'j.n.Y H:i' )
                 ->set_return_format( 'Y-m-d H:i:s' )
                 ->redipress_add_queryable( 'end_datetime' );
+
+            $recurring_event = ( new Field\TrueFalse( $strings['recurring_event']['label'] ) )
+                ->set_key( "{$key}_recurring_event" )
+                ->set_name( 'recurring_event' )
+                ->set_instructions( $strings['recurring_event']['instructions'] )
+                ->use_ui();
+
+            $recurring_dates = ( new Field\Repeater( $strings['recurring_dates']['label'] ) )
+                ->set_key( "{$key}_dates" )
+                ->set_name( 'dates' )
+                ->set_instructions( $strings['recurring_dates']['instructions'] )
+                ->set_button_label( $strings['recurring_dates']['button_label'] );
+
+            $recurring_start_datetime = ( new Field\DateTimePicker( $strings['start_datetime']['label'] ) )
+                ->set_key( "{$key}_start" )
+                ->set_name( 'start' )
+                ->set_instructions( $strings['start_datetime']['instructions'] )
+                ->set_display_format( 'j.n.Y H:i' )
+                ->set_return_format( 'Y-m-d H:i:s' );
+
+            $recurring_end_datetime = ( new Field\DateTimePicker( $strings['end_datetime']['label'] ) )
+                ->set_key( "{$key}_end" )
+                ->set_name( 'end' )
+                ->set_instructions( $strings['end_datetime']['instructions'] )
+                ->set_display_format( 'j.n.Y H:i' )
+                ->set_return_format( 'Y-m-d H:i:s' );
 
             $location_name = ( new Field\Text( $strings['location']['name']['label'] ) )
                 ->set_key( "{$key}_location_name" )
@@ -430,15 +465,25 @@ class ManualEvent {
                 ->add_rule( $price_is_free->get_key(), '!=', '1' );
             $rule_group_is_virtual_event = ( new ConditionalLogicGroup() )
                 ->add_rule( $is_virtual_event->get_key(), '==', '1' );
+            $rule_group_is_not_recurring = ( new ConditionalLogicGroup() )
+                ->add_rule( $recurring_event->get_key(), '==', '0' );
+            $rule_group_is_recurring     = ( new ConditionalLogicGroup() )
+                ->add_rule( $recurring_event->get_key(), '==', '1' );
 
             $price_group->add_conditional_logic( $rule_group_has_price );
             $virtual_event_link->add_conditional_logic( $rule_group_is_virtual_event );
+            $start_datetime->add_conditional_logic( $rule_group_is_not_recurring );
+            $end_datetime->add_conditional_logic( $rule_group_is_not_recurring );
+            $recurring_dates->add_conditional_logic( $rule_group_is_recurring );
+            $recurring_dates->add_fields( [ $recurring_start_datetime, $recurring_end_datetime ] );
 
             $tab->add_fields( [
                 $description,
                 $short_description,
+                $recurring_event,
                 $start_datetime,
                 $end_datetime,
+                $recurring_dates,
                 $location_group,
                 $price_is_free,
                 $price_group,
@@ -482,6 +527,8 @@ class ManualEvent {
             'location_title'     => __( 'Location', 'tms-theme-base' ),
             'price_title'        => __( 'Price', 'tms-theme-base' ),
             'provider_title'     => __( 'Organizer', 'tms-theme-base' ),
+            'recurring'          => ! empty( $event->dates ) ? count( $event->dates ) > 1 : null,
+            'dates'              => static::get_event_dates( $event ),
         ];
 
         if ( ! empty( $event->location ) ) {
@@ -614,5 +661,60 @@ class ManualEvent {
         }
 
         return null;
+    }
+
+    /**
+     * Get event dates info
+     *
+     * @param object $event Event object.
+     *
+     * @return array
+     */
+    public static function get_event_dates( $event ) {
+        $dates = [];
+
+        if ( empty( $event->dates ) ) {
+            return $dates;
+        }
+
+        foreach ( $event->dates as $date ) {
+            $dates[] = [
+                'date' => self::compare_dates( $date['start'], $date['end'] ),
+            ];
+        }
+
+        return $dates;
+    }
+
+    /**
+     * Get event date
+     *
+     * @param string $start Event startdate.
+     * @param string $end Event enddate.
+     *
+     * @return string|null
+     */
+    public static function compare_dates( $start, $end ) {
+        if ( empty( $start ) ) {
+            return null;
+        }
+
+        $start_time  = static::get_as_datetime( $start );
+        $end_time    = static::get_as_datetime( $end );
+        $date_format = 'j.n.Y H.i';
+
+        if ( $start_time && $end_time && $start_time->diff( $end_time )->days >= 1 ) {
+            return sprintf(
+                '%s - %s',
+                $start_time->format( $date_format ),
+                $end_time->format( $date_format )
+            );
+        }
+
+        return sprintf(
+            '%s - %s',
+            $start_time->format( 'j.n.Y H.i' ),
+            $end_time->format( 'H.i' )
+        );
     }
 }